Applications/PolynomialDiscovery/requirements.txt
academy-py
matplotlib
numpy
